Refactor: recoil을 관리하는 폴더를 store로 변경한다
