+++ src/components/common/ModalViewer/ModalViewer.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useRecoilState } from 'recoil';
-import { $modalState, ModalKey } from '@/recoil';
+import { $modalState, ModalKey } from '@/store';
 import { AlertModalDialog } from '@/components';
 
 const ModalViewer = () => {
