+++ src/store/modal.ts
